+++ app/src/androidTest/java/com/rrajath/grove/ui/editor/EditNoteScreenTest.kt
@@ -2,7 +2,7 @@ package com.rrajath.grove.ui.editor
 
 import androidx.activity.ComponentActivity
 import androidx.compose.ui.test.assertTextContains
-import androidx.compose.ui.test.junit4.createAndroidComposeRule
+import androidx.compose.ui.test.junit4.v2.createAndroidComposeRule
 import androidx.compose.ui.test.onAllNodesWithContentDescription
 import androidx.compose.ui.test.onAllNodesWithText
 import androidx.compose.ui.test.onNodeWithContentDescription