+++ app/src/androidTest/java/com/rrajath/grove/ui/screens/OnboardingScreenTest.kt
@@ -5,7 +5,7 @@ import android.app.Instrumentation.ActivityResult
 import android.content.Intent
 import androidx.activity.ComponentActivity
 import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.junit4.createAndroidComposeRule
+import androidx.compose.ui.test.junit4.v2.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performScrollTo